add force clear and refresh local
force clear is useful when going to playstate and leaving playstate to a menu since you won't be sharing any assets and it prevents the playstate assets from staying loaded. local refresh is useful for editor states since you can force the playstate assets to stay loaded even when they wouldn't

+++ source/transition/BaseTransition.hx
@@ -43,7 +43,13 @@ class BaseTransition extends FlxSpriteGroup{
 		if(state != null){ //State exit animation.
 			//FlxG.signals.postStateSwitch.addOnce(Utils.gc);
 			FlxG.signals.preStateCreate.addOnce(function(state){
-				ImageCache.destroyByCount();
+				if(!ImageCache.forceClearOnTransition){
+					ImageCache.destroyByCount();
+				}
+				else{
+					ImageCache.clearLocal();
+					ImageCache.forceClearOnTransition = false;
+				}
 				AudioCache.clear();
 				
 				Utils.gc();
